fix ldm issue

+++ src/migrations/20231030150724-sim_score_cache.js
@@ -20,7 +20,7 @@ module.exports = {
           recipient_id: { allowNull: false, type: Sequelize.INTEGER },
           goal1: { allowNull: false, type: Sequelize.INTEGER },
           goal2: { allowNull: false, type: Sequelize.INTEGER },
-          score: { allowNull: false, type: Sequelize.FLOAT },
+          score: { allowNull: false, type: Sequelize.DOUBLE },
           createdAt: { allowNull: false, type: Sequelize.DATE },
           updatedAt: { allowNull: false, type: Sequelize.DATE },
         },
